initial commit: create a csv transaction processor
The code included here can parse a csv file with a known format and
execute transactions on ledger based off of that. These transactions
are: deposit, withdrawal, dispute, resolve and chargeback.
All transactions are implemented, but there are known issues (like
overflows which can cause panics) with it as it is.