transactions: add overflow and underflow checks for disputes
Disputing a transaction can result in an overflow or underflow.
For instance, consider the case where you:

1: deposit 10
2: withdraw 10
3: dispute 1

The dispute operation will underflow as we attempt to remove available
funds that were already withdrawn.

Another possible errors include:
* held values by disputing many transactions
* overflowing available funds by resolving a dispute after a large deposit

Note that chargebacks can't cause underflows: Since we only chargeback a
transaction marked as disputed we are guaranteed to have these funds
available when we subtract them from the account.