account: move deposit and withdrawal functions into account
in order to streamline our code and consolidate DisputeEntry and
TransactionEntry into a single type, let's start to move the parts of
the code which change data internal to account into account itself.
We do it first with deposit and withdrawal as the logic around these
methods is relatively simple and they only require an immutable "amount"
as a parameter.