accounts: change number representation to a more precise one
when adding a test to check the precision of the fixed crate, we noticed
that it was much lower then we expected. Switch to rust_decimal instead
with seems to not introduce too much of a performance penalty compared
to fixed. It can also represent much larger numbers then the previous
crate.