account: be more lenient with locked accounts
in hindsight, locked accounts should be able to do any operation other
then withdrawal. The fact that one chargeback occurred does not mean
that other chargebacks won't occur, or that the client shouldn't be able
to deposit more money into the account. Let's also not lock the account
unless a chargeback happens since that's what the specification says.